[SPARK-40772][SQL] Improve `spark.sql.adaptive.skewJoin.skewedPartitionFactor` to support `Double` values (#1591)
### What changes were proposed in this pull request?

This PR aims to improve `spark.sql.adaptive.skewJoin.skewedPartitionFactor` to support float values by converging to `doubleConf` from `intConf`.

### Why are the changes needed?

Like `spark.sql.adaptive.rebalancePartitionsSmallPartitionFactor`, this allows users to use the configuration more flexibly.

### Does this PR introduce _any_ user-facing change?

Yes, but it will accept all previous Integer configuration values.

### How was this patch tested?

Pass the CIs with the changed default value, `5.0`.

Authored-by: Dongjoon Hyun <dongjoon@apache.org>
(cherry picked from commit 8251f876fd6a0300bbdac415091786dda3957532)
(cherry picked from commit fe9fab2b06a49b039325166521b1eb09bafd75c1)

+++ sql/core/src/main/scala/org/apache/spark/sql/execution/adaptive/OptimizeSkewedJoin.scala
@@ -59,7 +59,7 @@ object OptimizeSkewedJoin extends AQEShuffleReadRule {
    */
   def getSkewThreshold(medianSize: Long): Long = {
     conf.getConf(SQLConf.SKEW_JOIN_SKEWED_PARTITION_THRESHOLD).max(
-      medianSize * conf.getConf(SQLConf.SKEW_JOIN_SKEWED_PARTITION_FACTOR))
+      (medianSize * conf.getConf(SQLConf.SKEW_JOIN_SKEWED_PARTITION_FACTOR)).toLong)
   }
 
   private def medianSize(sizes: Array[Long]): Long = {
